use Asciidoctor to build documentation when possible
The Asciidoctor project is more actively maintained than asciidoc.  Use
it for building the documentation on Fedora.  Asciidoctor is not
currently available for EL-6 or EL-8, though it is in EPEL for EL-7.
Exclude all EL builds for now, until we can reliably use it on EL-7 and
EL-8 (including CentOS-Stream, ideally).

This is made possible by the excellent work of both the Git and
Asciidoctor communities.  Thanks in particular to brian m. carlson,
Martin Ågren, Jeff King, and Dan Allen.