adjust make test options
The make test call was changed to use %make_build in d34bc42 (Use
make_build macro when running tests, 2020-01-14) in order to allow the
options to be more easily overridden.  This enabled the -O option by
default, which causes the test output to be printed only after all the
tests have run.

That makes following the progress in both interactive and copr/koji
builds difficult.  Replace %make_build with %__make to drop the unwanted
-O option but still allow the make command to be overridden.