fix/quiet rpmlint issues from libsecret split
When the libsecret credential helper was split out in 9d91bab (split
libsecret credential helper into a subpackage (#1804741), 2020-02-19), a
few rpmlint errors & warnings crept in.

Update the rpmlintrc file to ignore the no-documentation warning for the
libsecret subpackage (replacing the gnome-keyring entry which is no
longer needed).  Fix an errant tab added to the spec file.

Moving the libsecret credential helper to a subpackage left no binaries
in the main git package, so rpmlint complains.  Fixing this requires a
bit more investigation and care.