adjust python hashbang in contrib/hg-to-git, it supports python3
The hg-to-git.py script in contrib grew python3 support in upstream
commit d17ae00c97 (hg-to-git: make it compatible with both python3 and
python2, 2019-09-18), which was released in git-2.24.0.  Move it from
the python2-only conditionals.

(This leaves contrib/fast-import/import-zips.py as the sole python
script which is _not_ python3-compatible.  It seems to need only minimal
fixes for python2/python3 compatibility -- per some light testing.)