move git-difftool to git-core, it does not require perl
The `git difftool` command was converted to a builtin in git-2.12.0
(from 2017).  We don't need to split it out of git-core.

This was missed in cb7fab7 (Move commands which no longer require perl
into git-core, 2017-11-10) and d56cfc6 (Use symlinks instead of
hardlinks for installed binaries, 2018-03-15).  Better late than never.