update Junio's GPG key (with extended expiration)
The key used to sign git releases expired in July 2020.  While this
doesn't strictly affect us because use gpgv to verify the releases
against a known key file, it is worth updating to make it clear that
we're using the correct signing key.

Refer to 7c95c76 (Update Junio's GPG key, 2017-09-16) for a previous
update of the key, including the process used.

Here is a diff of the key file before and after the update:

    $ diff -u <(gpg gpgkey-junio.asc.old 2>/dev/null) <(gpg gpgkey-junio.asc 2>/dev/null)
    --- /dev/fd/63	2021-01-25 11:57:17.367151191 -0500
    +++ /dev/fd/62	2021-01-25 11:57:17.368151229 -0500
    @@ -3,6 +3,6 @@
     uid                           Junio C Hamano <gitster@pobox.com>
     uid                           Junio C Hamano <junio@pobox.com>
     uid                           Junio C Hamano <jch@google.com>
    -sub   rsa4096/B0B5E88696AFE6CB 2011-10-03 [S] [expired: 2020-07-26]
    +sub   rsa4096/B0B5E88696AFE6CB 2011-10-03 [S] [expires: 2028-01-11]
     sub   rsa4096/86B76D5D833262C4 2011-10-01 [E]
    -sub   rsa4096/7594EEC7B3F7CAC9 2014-09-20 [S] [expired: 2020-07-26]
    +sub   rsa4096/7594EEC7B3F7CAC9 2014-09-20 [S] [expires: 2028-01-11]

This thread on the git list is where the question was raised and Junio
confirmed he'd extended the expiration of his signing key:

https://lore.kernel.org/git/B6DFB74D-A722-4DBD-A4B2-562604B21CCB@alchemists.io/T/#u